Accept POST requests on /v1/submit_survey endpoint
Adds filters and handler to support POST requests to /v1/submit_survey endpoint like
{
    device_id: Uuid (string),
    submitted: Optional timestamp (utc),
    survey_data: String
}
survey_data will likely change. A single string is not a useful means for storing survey submissions,
but I need to find the syntax for storing a JSON blob in Diesel.

Alternatively each question/answer pair could be stored as its own field, but this
requires a database change for every survey change which does not feel right. The survey
should be able to change without requiring a server rollout. This has forward/backward
compatibility risks. Storing the survey as a JSON blob mitigates the API versioning fiasco.